more cyclic imports cleanup
